+++ Octokit.Reactive/Clients/ObservableReleasesClient.cs
@@ -100,7 +100,7 @@ namespace Octokit.Reactive
         /// <param name="name">The repository's name</param>
         /// <param name="id">The id of the release</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
-        public IObservable<Release> Get(string owner, string name, int id)
+        public IObservable<Release> Get(string owner, string name, long id)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -136,7 +136,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="id">The id of the release</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
-        public IObservable<Release> Get(long repositoryId, int id)
+        public IObservable<Release> Get(long repositoryId, long id)
         {
             return _client.Get(repositoryId, id).ToObservable();
         }
@@ -233,7 +233,7 @@ namespace Octokit.Reactive
         /// <param name="id">The id of the release</param>
         /// <param name="data">A description of the release to edit</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
-        public IObservable<Release> Edit(string owner, string name, int id, ReleaseUpdate data)
+        public IObservable<Release> Edit(string owner, string name, long id, ReleaseUpdate data)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -252,7 +252,7 @@ namespace Octokit.Reactive
         /// <param name="id">The id of the release</param>
         /// <param name="data">A description of the release to edit</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
-        public IObservable<Release> Edit(long repositoryId, int id, ReleaseUpdate data)
+        public IObservable<Release> Edit(long repositoryId, long id, ReleaseUpdate data)
         {
             Ensure.ArgumentNotNull(data, nameof(data));
 
@@ -269,7 +269,7 @@ namespace Octokit.Reactive
         /// <param name="name">The repository's name</param>
         /// <param name="id">The id of the release to delete</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
-        public IObservable<Unit> Delete(string owner, string name, int id)
+        public IObservable<Unit> Delete(string owner, string name, long id)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -286,7 +286,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="id">The id of the release to delete</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
-        public IObservable<Unit> Delete(long repositoryId, int id)
+        public IObservable<Unit> Delete(long repositoryId, long id)
         {
             return _client.Delete(repositoryId, id).ToObservable();
         }
@@ -301,7 +301,7 @@ namespace Octokit.Reactive
         /// <param name="name">The repository's name</param>
         /// <param name="id">The id of the <see cref="Release"/>.</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
-        public IObservable<ReleaseAsset> GetAllAssets(string owner, string name, int id)
+        public IObservable<ReleaseAsset> GetAllAssets(string owner, string name, long id)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -318,7 +318,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="id">The id of the <see cref="Release"/>.</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
-        public IObservable<ReleaseAsset> GetAllAssets(long repositoryId, int id)
+        public IObservable<ReleaseAsset> GetAllAssets(long repositoryId, long id)
         {
             return GetAllAssets(repositoryId, id, ApiOptions.None);
         }
@@ -334,7 +334,7 @@ namespace Octokit.Reactive
         /// <param name="id">The id of the <see cref="Release"/>.</param>
         /// <param name="options">Options for changing the API response</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
-        public IObservable<ReleaseAsset> GetAllAssets(string owner, string name, int id, ApiOptions options)
+        public IObservable<ReleaseAsset> GetAllAssets(string owner, string name, long id, ApiOptions options)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -353,7 +353,7 @@ namespace Octokit.Reactive
         /// <param name="id">The id of the <see cref="Release"/>.</param>
         /// <param name="options">Options for changing the API response</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
-        public IObservable<ReleaseAsset> GetAllAssets(long repositoryId, int id, ApiOptions options)
+        public IObservable<ReleaseAsset> GetAllAssets(long repositoryId, long id, ApiOptions options)
         {
             Ensure.ArgumentNotNull(options, nameof(options));
 
@@ -369,7 +369,7 @@ namespace Octokit.Reactive
         /// <param name="owner">The repository's owner</param>
         /// <param name="name">The repository's name</param>
         /// <param name="assetId">The id of the <see cref="ReleaseAsset"/></param>
-        public IObservable<ReleaseAsset> GetAsset(string owner, string name, int assetId)
+        public IObservable<ReleaseAsset> GetAsset(string owner, string name, long assetId)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -386,7 +386,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="assetId">The id of the <see cref="ReleaseAsset"/></param>
-        public IObservable<ReleaseAsset> GetAsset(long repositoryId, int assetId)
+        public IObservable<ReleaseAsset> GetAsset(long repositoryId, long assetId)
         {
             Ensure.ArgumentNotNull(assetId, nameof(assetId));
 
@@ -421,7 +421,7 @@ namespace Octokit.Reactive
         /// <param name="name">The repository's name</param>
         /// <param name="assetId">The id of the <see cref="ReleaseAsset"/></param>
         /// <param name="data">Description of the asset with its amended data</param>
-        public IObservable<ReleaseAsset> EditAsset(string owner, string name, int assetId, ReleaseAssetUpdate data)
+        public IObservable<ReleaseAsset> EditAsset(string owner, string name, long assetId, ReleaseAssetUpdate data)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -439,7 +439,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="assetId">The id of the <see cref="ReleaseAsset"/></param>
         /// <param name="data">Description of the asset with its amended data</param>
-        public IObservable<ReleaseAsset> EditAsset(long repositoryId, int assetId, ReleaseAssetUpdate data)
+        public IObservable<ReleaseAsset> EditAsset(long repositoryId, long assetId, ReleaseAssetUpdate data)
         {
             Ensure.ArgumentNotNull(data, nameof(data));
 
@@ -455,7 +455,7 @@ namespace Octokit.Reactive
         /// <param name="owner">The repository's owner</param>
         /// <param name="name">The repository's name</param>
         /// <param name="id">The id of the <see cref="ReleaseAsset"/>.</param>
-        public IObservable<Unit> DeleteAsset(string owner, string name, int id)
+        public IObservable<Unit> DeleteAsset(string owner, string name, long id)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -471,7 +471,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="id">The id of the <see cref="ReleaseAsset"/>.</param>
-        public IObservable<Unit> DeleteAsset(long repositoryId, int id)
+        public IObservable<Unit> DeleteAsset(long repositoryId, long id)
         {
             return _client.DeleteAsset(repositoryId, id).ToObservable();
         }
